scanners/requirements.txt
python-nmap>=0.7.1
scapy>=2.5.0
zeroconf>=0.131.0
mac-vendor-lookup>=0.1.12
netifaces>=0.11.0
